app/requirements.txt
h2o-wave==0.19.0
python-arango==7.1.0